wordpress_requirements.txt
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
Pillow>=10.0.0
google-generativeai>=0.3.0
